==> notebooks/04-data_preparation.ipynb <==
## Data preparation for Steam datasets

import pandas as pd
import pickle
import string
import nltk
import re      
import ast
from nltk.stem.porter import PorterStemmer
from nltk.stem import WordNetLemmatizer
from nltk.tokenize import word_tokenize
from sklearn.preprocessing import MultiLabelBinarizer

### PLAYTIME DATA PREP

# load steam_playtime.csv
users_df = pd.read_csv('../data/steam_playtime.csv')

len(users_df)

user playtimes inliers max/min values based on z-score and iqr as shown in data analysis<br>
z_max = ~128000<br>
z_min = 1<br>
iqr_max = ~1600<br>
iqr_min = 1

# drop counterstrike i.e. drop rows where appid is 730, 10 or 240
# users_df = users_df[~users_df['appid'].isin([730, 10, 240])]

# remove user playtimes over x minutes
users_df = users_df[users_df['playtime_forever'] < 1700]

# remove user playtimes under x minutes
#users_df = users_df[users_df['playtime_forever'] > 10]

# RUN THIS IF YOU WANT TO DROP GAMES WITH ONLY ONE PLAYTIME
# group by appid and get count, mean, median, min, max
playtime_by_game = users_df.groupby('appid').agg({'playtime_forever': ['count', 'mean', 'median', 'min', 'max']}).round(2)
# get appids where count is 1
only_one_playtime = playtime_by_game[playtime_by_game[('playtime_forever', 'count')] == 1].index
# exclude games with only one playtime
users_df = users_df[~users_df['appid'].isin(only_one_playtime)]

### GAME DATA PREP

# load steam_app_metadata.csv
games_df = pd.read_csv('../data/steam_app_metadata.csv')

# remove duplicate appid
games_df = games_df.drop_duplicates(subset=['appid'])

# get duplicates for name
dupl_game_names = games_df[games_df.duplicated(subset=['name'])]

# rename appid's in users_df that are duplicates to original appid
for index, row in dupl_game_names.iterrows():
    dupl_appid = row['appid']
    orig_appid = games_df[games_df['name'] == row['name']]['appid'].iloc[0]
    users_df.loc[users_df['appid'] == dupl_appid, 'appid'] = orig_appid

# remove duplicate game names from games_df
games_df = games_df.drop_duplicates(subset=['name'])

no_description = games_df[games_df['description'].isnull() | (games_df['description'] == '')]
games_df = games_df[~games_df['appid'].isin(no_description['appid'])]

# remove games with no publisher or developer
no_dev_or_pub = games_df[(games_df['developer'] == 'None') & games_df['publisher'].isnull()]
games_df = games_df[~games_df['appid'].isin(no_dev_or_pub['appid'])]

# Fill NaN for publisher with developer
games_df['publisher'] = games_df['publisher'].fillna(games_df['developer'])

# remove games from games_df that are not in users_df
games_df = games_df[games_df['appid'].isin(users_df['appid'].unique())]

# create games df for flask app
games_flask_df = games_df[['appid', 'name', 'developer', 'publisher', 'description']]

# prepend developer to description_clean
games_df['description'] = games_df['developer'] + ' ' + games_df['description']

#### one hot encode categories and genres

# make categories and genres into lists
games_df['categories'] = games_df['categories'].apply(ast.literal_eval)
games_df['genres'] = games_df['genres'].apply(ast.literal_eval)

# games_flask_df genres and categories lists as comma separated strings into games_flask_df
games_flask_df['categories'] = games_df['categories'].apply(lambda x: ', '.join(x))
games_flask_df['genres'] = games_df['genres'].apply(lambda x: ', '.join(x))

# multilabel binarizer
mlb = MultiLabelBinarizer()

# encode categories 
games_df = games_df.join(pd.DataFrame(mlb.fit_transform(games_df.pop('categories')),
                          columns=mlb.classes_,
                          index=games_df.index))

# encode genres
games_df = games_df.join(pd.DataFrame(mlb.fit_transform(games_df.pop('genres')),
                          columns=mlb.classes_,
                          index=games_df.index))

#### process description column NLP style

nltk.download('wordnet')
stopwords = nltk.corpus.stopwords.words('english')

# lemmatizer and stemmer. Only use one of them
# porter = PorterStemmer()
lemmatizer = WordNetLemmatizer()

tag_regex = re.compile(r'<[^>]+>')

# function to clean text for flask app
def clean_text_for_app(text):
    # remove tags
    text = tag_regex.sub(' ', text)
    # remove URL links
    text = re.sub(r'http\S+', ' ', text)
    # remove hex
    text = re.sub(r'[^\x00-\x7f]',r'', text)
    return text


# function to clean text
def clean_text(text):
    # remove tags
    text = tag_regex.sub(' ', text)
    # remove URL links
    text = re.sub(r'http\S+', ' ', text)
    # remove non-breaking space
    text = text.replace(u'\xa0', u' ')
    # remove hex
    text = re.sub(r'[^\x00-\x7f]',r'', text)
    # hyphens to space
    text = text.replace('-', ' ')
    # remove "&quot;
    text = text.replace('&quot;', '')
    # remove punctuation
    text = text.translate(str.maketrans('', '', string.punctuation))
    # remove numbers
    text = text.translate(str.maketrans('', '', string.digits))
    # make lowercase
    text = text.lower()
    # tokenize text
    tokens = word_tokenize(text)
    # remove stopwords
    tokens = [w for w in tokens if w not in stopwords]
    # lemmatize words
    tokens = [lemmatizer.lemmatize(w) for w in tokens]
    # stem words
    # tokens = [porter.stem(w) for w in tokens]
    # limit to 140 tokens
    tokens = tokens[:140]
    # make tokens into string
    text = ' '.join(tokens)
    # return clean text
    return text

# apply clean_text function to description column
games_df['description_clean'] = games_df['description'].apply(clean_text)

games_flask_df['description'] = games_flask_df['description'].apply(clean_text_for_app)

#### Save some dataframes and other stuff

users_df = users_df.drop(columns= ['playtime_2weeks'])

# drop rows from users_df where appid is not in games_df
# playtime for games is still in user profiles even if a game is no longer on steam
users_df = users_df[users_df['appid'].isin(games_df['appid'])]

# check datasets appid length
len(users_df['appid'].unique()) == len(games_df['appid'].unique() == len(games_df))

# save users_df to csv
users_df.to_csv('../data/steam_playtime_clean.csv', index=False)

# drop features not needed for model
games_df = games_df.drop(columns = ['developer', 'publisher', 'description'])

# reset games_df index
games_df = games_df.reset_index(drop=True)

# save games_df to pickle
games_df.to_pickle('../data/steam_games_clean.pkl')

games_df['description_clean']

# average number of words in description_clean, it was around 140 before limiting
games_df['description_clean'].apply(lambda x: len(x.split(' '))).mean()

# save a version for flask_app
games_flask_df.to_csv('../data/steam_metadata_flask.csv', index=False)
